chore: Split custom_fee.py into three separate examples
